requirements.txt
asgiref==3.6.0
Django==4.0.10
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-yasg==1.21.7
gunicorn==20.1.0
inflection==0.5.1
packaging==23.2
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.3
uritemplate==4.1.1
whitenoise==6.0.0
